Adding pseudo-polar intermediate representation
The vertex and edge calculations are already done in terms of this representation, so factoring it out makes those a
little more natural and will allow us to reuse some of the math to project cartesian coordinates to faces. It does
come at the cost of some repeated calculation for clarity.